Move TaskDeferral to threading namespace
